+++ extension/src/test/e2e/extension.test.ts
@@ -1,64 +1,13 @@
 import { suite, before, describe, it } from 'mocha'
-import { ViewControl } from 'wdio-vscode-service'
+import {
+  closeAllEditors,
+  dismissAllNotifications,
+  getDVCActivityBarIcon,
+  waitForViewContainerToLoad
+} from './util'
 import { ExperimentsWebview } from './pageObjects/experimentsWebview'
 import { PlotsWebview } from './pageObjects/plotsWebview'
 import { delay } from '../../util/time'
-
-const getDVCActivityBarIcon = async (): Promise<ViewControl> => {
-  const workbench = await browser.getWorkbench()
-
-  const activityBar = workbench.getActivityBar()
-
-  await browser.waitUntil(
-    async () => !!(await activityBar.getViewControl('DVC'))
-  )
-  return activityBar.getViewControl('DVC') as Promise<ViewControl>
-}
-
-const dismissAllNotifications = () =>
-  browser.waitUntil(async () => {
-    const workbench = await browser.getWorkbench()
-    const notifications = await workbench.getNotifications()
-    for (const n of notifications) {
-      await n.dismiss()
-    }
-    const openNotifications = await workbench.getNotifications()
-    return openNotifications.length === 0
-  })
-
-const waitForViewContainerToLoad = async () => {
-  const initialProgressBars = await $$('.monaco-progress-container')
-  await browser.waitUntil(async () => {
-    const dvcIcon = await getDVCActivityBarIcon()
-    if (!dvcIcon) {
-      return false
-    }
-
-    const view = await dvcIcon.openView()
-
-    return !!view
-  })
-
-  return browser.waitUntil(async () => {
-    const numberOfProgressBarsInContainer = 7
-    const currentProgressBars = await $$('.monaco-progress-container')
-
-    if (
-      currentProgressBars.length <
-      initialProgressBars.length + numberOfProgressBarsInContainer
-    ) {
-      return false
-    }
-
-    for (const progress of currentProgressBars) {
-      if ((await progress.getAttribute('aria-hidden')) !== 'true') {
-        return false
-      }
-    }
-
-    return true
-  })
-}
 
 suite('DVC Extension For Visual Studio Code', () => {
   before('should finish loading the extension', async () => {
@@ -129,9 +78,7 @@ suite('DVC Extension For Visual Studio Code', () => {
 
   describe('Plots Webview', () => {
     before(async () => {
-      const workbench = await browser.getWorkbench()
-      const editorView = workbench.getEditorView()
-      return editorView.closeAllEditors()
+      await closeAllEditors()
     })
 
     const webview = new PlotsWebview('plots')
