+++ extension/src/test/e2e/util.ts
@@ -1,0 +1,63 @@
+import { ViewControl } from 'wdio-vscode-service'
+
+export const dismissAllNotifications = () =>
+  browser.waitUntil(async () => {
+    const workbench = await browser.getWorkbench()
+    const notifications = await workbench.getNotifications()
+    for (const n of notifications) {
+      await n.dismiss()
+    }
+    const openNotifications = await workbench.getNotifications()
+    return openNotifications.length === 0
+  })
+
+export const getDVCActivityBarIcon = async (): Promise<ViewControl> => {
+  const workbench = await browser.getWorkbench()
+
+  const activityBar = workbench.getActivityBar()
+
+  await browser.waitUntil(
+    async () => !!(await activityBar.getViewControl('DVC'))
+  )
+  return activityBar.getViewControl('DVC') as Promise<ViewControl>
+}
+
+export const waitForViewContainerToLoad = async () => {
+  const initialProgressBars = await $$('.monaco-progress-container')
+  await browser.waitUntil(async () => {
+    const dvcIcon = await getDVCActivityBarIcon()
+    if (!dvcIcon) {
+      return false
+    }
+
+    const view = await dvcIcon.openView()
+
+    return !!view
+  })
+
+  return browser.waitUntil(async () => {
+    const numberOfProgressBarsInContainer = 7
+    const currentProgressBars = await $$('.monaco-progress-container')
+
+    if (
+      currentProgressBars.length <
+      initialProgressBars.length + numberOfProgressBarsInContainer
+    ) {
+      return false
+    }
+
+    for (const progress of currentProgressBars) {
+      if ((await progress.getAttribute('aria-hidden')) !== 'true') {
+        return false
+      }
+    }
+
+    return true
+  })
+}
+
+export const closeAllEditors = async () => {
+  const workbench = await browser.getWorkbench()
+  const editorView = workbench.getEditorView()
+  return editorView.closeAllEditors()
+}
